// src/src/app/tramites/cancilleria/tramitar-pasaporte/models/dataSolicitudPasaporteModel.ts
export interface DataSolicitudPasaporteModel {    
    motivoSolicitud: string;
    oficina: string;
    tipoPasaporte: string;
    digitoVerificacionOCR: string;
    fechaExpedicionPasaporte: string;
    numeroPasaporte: string;
    codigoDocumento: string;
    correoElectronico: string;
    fechaExpedicionDocIdentidad: string;
    numeroPersonal: string;
}